updated plot1.R

+++ plot1.R
@@ -1,15 +1,19 @@
-## Script to create plot1.png
-## This was created as part of Coursera Data Science Specialization
-## Course: Exploratory Data Analysis, Course Project # 1
-## Author: Brent Brewington
+## Coursera - Exploratory Data Analysis
+## Course Project # 1
+## Author: Brent Brewington, (github: bbrewington)
+## Plot1.R
 
 # Get data from file "household_power_consumption", and save to data frame "DF"
-
-temp <- tempfile()
-download.file("https://d396qusza40orc.cloudfront.net/exdata%2Fdata%2Fhousehold_power_consumption.zip",temp)
-DF <- read.table(unz(temp, "household_power_consumption.txt"), 
+if(!("household_power_consumption.txt" %in% list.files())){
+  temp <- tempfile()
+  download.file("https://d396qusza40orc.cloudfront.net/exdata%2Fdata%2Fhousehold_power_consumption.zip",temp)
+  DF <- read.table(unz(temp, "household_power_consumption.txt"), 
                    header=TRUE,sep=";",na.strings="?",stringsAsFactors=FALSE)
-unlink(temp)
+  unlink(temp)
+} else{
+  DF <- read.table("household_power_consumption.txt", 
+                   header=TRUE,sep=";",na.strings="?",stringsAsFactors=FALSE)
+}
 
 # Convert DF$Date to POSIXlt and save to new column called "Date_formatted"
 
@@ -22,6 +26,6 @@ DF_subset <- subset(DF, Date_formatted == "2007-02-01" | Date_formatted == "2007
 
 # Create histogram and save to png file "plot1.png" in the working directory
 png(file="plot1.png", width=480, height=480)
-hist(DF_subset$Global_active_power, xaxp=c(0,6,3),yaxp=c(0,1200,3), main="Global Active Power", 
+hist(DF_subset$Global_active_power, xaxp=c(0,6,3),yaxp=c(0,1200,6), main="Global Active Power", 
      xlab = "Global Active Power (kilowatts)", ylab = "Frequency", col = "Red")
 dev.off()
